fixup! Fix permissive errors on GCC 4.7.2 build

+++ Release/src/json/json.cpp
@@ -288,7 +288,7 @@ bool web::json::number::is_int32() const
         case signed_type:
             return m_intval >= (std::numeric_limits<int32_t>::min)() &&
                    m_intval <= (std::numeric_limits<int32_t>::max)();
-        case unsigned_type: return m_uintval <= (std::numeric_limits<int32_t>::max)();
+        case unsigned_type: return m_uintval <= (std::numeric_limits<uint32_t>::max)();
         case double_type:
         default: return false;
     }
@@ -298,8 +298,8 @@ bool web::json::number::is_uint32() const
 {
     switch (m_type)
     {
-        case signed_type: return m_intval >= 0 && m_intval <= (std::numeric_limits<uint32_t>::max)();
-        case unsigned_type: return m_uintval <= (unsigned)(std::numeric_limits<uint32_t>::max)();
+        case signed_type: return m_intval >= 0 && m_intval <= (std::numeric_limits<int32_t>::max)();
+        case unsigned_type: return m_uintval <= (std::numeric_limits<uint32_t>::max)();
         case double_type:
         default: return false;
     }
